feat(graph): Update data flow for graph rendering
- Refactor how graph data is processed and passed to React Flow by
  centralizing the logic in `graphToReactFlow` function. This
  reduces code duplication and enhances clarity around graph data
  structure.
- Introduce logging to assist with debugging graph properties
  during development.
