+++ app/(playground)/p/[agentId]/beta-proto/graph/provider.tsx
@@ -53,6 +53,7 @@ export const GraphProvider: FC<PropsWithChildren<GraphProviderProps>> = ({
 			deboucedSetGraphToDb(stateRef.current.graph);
 		}
 	}, [deboucedSetGraphToDb]);
+	console.log({ defaultGraph });
 	return (
 		<GraphContext.Provider value={{ state, dispatch: enhancedDispatch }}>
 			{children}
