+++ app/(playground)/p/[agentId]/beta-proto/react-flow-adapter/graph.tsx
@@ -29,42 +29,7 @@ import type { Graph } from "../graph/types";
 import { removeConnector } from "../graph/v2/composition/remove-connector";
 import { removeNode } from "../graph/v2/composition/remove-node";
 import { setXyFlowEdges, setXyFlowNodes } from "../graph/v2/xy-flow";
-import {
-	type ReactFlowEdge,
-	type ReactFlowNode,
-	giselleEdgeType,
-	giselleNodeType,
-} from "./giselle-node";
-
-export function graphToReactFlow(grpah: Graph) {
-	const nodes: ReactFlowNode[] = grpah.nodes.map((node) => {
-		return {
-			id: node.id,
-			type: giselleNodeType,
-			position: node.ui.position,
-			selected: node.ui.selected,
-			data: {
-				...node,
-			},
-		};
-	});
-
-	const edges: ReactFlowEdge[] = grpah.connectors.map((connector) => {
-		return {
-			id: connector.id,
-			type: giselleEdgeType,
-			source: connector.source,
-			target: connector.target,
-			targetHandle: connector.targetHandle,
-			data: connector,
-		};
-	});
-
-	return {
-		nodes,
-		edges,
-	};
-}
+import type { ReactFlowEdge, ReactFlowNode } from "./types";
 
 export const useReactFlowNodeEventHandler = () => {
 	const { state, dispatch } = useGraph();
@@ -143,43 +108,6 @@ export function useReacrFlowEdgeEventHandler() {
 
 	return { handleEdgesChange };
 }
-
-export const useGraphToReactFlowEffect = () => {
-	const { state, dispatch } = useGraph();
-	const reactFlowInstance = useReactFlow();
-
-	useEffect(() => {
-		const { nodes, edges } = graphToReactFlow(state.graph);
-		reactFlowInstance.setNodes(nodes);
-		reactFlowInstance.setEdges(edges);
-	}, [reactFlowInstance.setNodes, reactFlowInstance.setEdges, state.graph]);
-
-	const onChange = useCallback<OnSelectionChangeFunc>(
-		({ nodes }) => {
-			if (nodes.length === 1) {
-				dispatch(
-					selectNodeAndSetPanelTab({
-						selectNode: {
-							id: nodes[0].id as GiselleNodeId,
-							panelTab: panelTabs.property,
-						},
-					}),
-				);
-			} else {
-				dispatch(
-					selectNode({
-						selectedNodeIds: nodes.map((node) => node.id as GiselleNodeId),
-					}),
-				);
-			}
-		},
-		[dispatch],
-	);
-
-	// useOnSelectionChange({
-	// 	onChange,
-	// });
-};
 
 type GiselleConnection = {
 	source: GiselleNodeId;
